requirements.txt
Django>=3.2.0
djangorestframework>=3.12.0
django-cors-headers>=3.10.0
PyPDF2>=3.0.0
